@@ -12,7 +12,7 @@ namespace Southwind.Load.NorthwindSchema
 {
     public static class Northwind
     {
-        static Connector connector;
+        static Connector? connector;
         public static Connector Connector
         {
             get
