@@ -30,12 +30,13 @@ using Signum.Entities.MachineLearning;
 using Signum.Engine.MachineLearning;
 using Microsoft.Extensions.Configuration;
 using System.IO;
+using System.Reflection;
 
 namespace Southwind.Load
 {
     class Program
     {
-        public static IConfigurationRoot ConfigRoot;
+        public static IConfigurationRoot ConfigRoot = null!;
 
         static int Main(string[] args)
         {
@@ -100,18 +101,28 @@ namespace Southwind.Load
             {
                 SafeConsole.WriteColor(ConsoleColor.DarkRed, e.GetType().Name + ": ");
                 SafeConsole.WriteLineColor(ConsoleColor.Red, e.Message);
-                SafeConsole.WriteLineColor(ConsoleColor.DarkRed, e.StackTrace.Indent(4));
+                SafeConsole.WriteLineColor(ConsoleColor.DarkRed, e.StackTrace!.Indent(4));
                 return -1;
             }
         }
 
+        private static void GetLite()
+        {
+            var juas = Database.Query<RoleEntity>().Single(a => a.Name == "Juas");
 
+            new RoleEntity
+            {
+                Name = "Bla",
+                Roles = { juas.ToLite()}
+            }.Save();
+        }
 
         private static void Load(string[]? args)
         {
             Schema.Current.Initialize();
 
             OperationLogic.AllowSaveGlobally = true;
+            GetLite();
 
             while (true)
             {
